🔄 Fix circular dependencies
