+++ .changeset/few-bulldogs-brush.md
@@ -1,0 +1,8 @@
+---
+'myst-frontmatter': patch
+'myst-parser': patch
+'tex-to-myst': patch
+'myst-cli': patch
+---
+
+Fix circular dependencies
