+++ packages/myst-cli/src/project/load.ts
@@ -3,11 +3,14 @@ import { join, resolve } from 'node:path';
 import { isDirectory, isUrl } from 'myst-cli-utils';
 import { RuleId } from 'myst-common';
 import { loadConfig } from '../config.js';
-import { loadFile, combineProjectCitationRenderers } from '../process/index.js';
+import { combineProjectCitationRenderers } from '../process/citations.js';
+import { loadFile } from '../process/file.js';
 import type { ISession } from '../session/types.js';
 import { selectors } from '../store/index.js';
 import { projects } from '../store/reducers.js';
-import { addWarningForFile, getAllBibTexFilesOnPath, validateTOC } from '../utils/index.js';
+import { addWarningForFile } from '../utils/addWarningForFile.js';
+import { getAllBibTexFilesOnPath } from '../utils/getAllBibtexFiles.js';
+import { validateTOC } from '../utils/toc.js';
 import { projectFromPath } from './fromPath.js';
 import { projectFromToc } from './fromToc.js';
 import { writeTocFromProject } from './toToc.js';
