+++ packages/myst-cli/src/transforms/embed.ts
@@ -4,7 +4,7 @@ import type { IReferenceState, MultiPageReferenceState } from 'myst-transforms';
 import type { GenericNode, GenericParent } from 'myst-common';
 import { copyNode, liftChildren, normalizeLabel } from 'myst-common';
 import type { Dependency, Embed, Container } from 'myst-spec-ext';
-import { selectFile } from '../process/index.js';
+import { selectFile } from '../process/file.js';
 import type { ISession } from '../session/types.js';
 
 /**
