+++ packages/myst-parser/src/index.ts
@@ -3,7 +3,6 @@ import * as plugins from './plugins.js';
 export * from './myst.js';
 export * from './types.js';
 export * from './fromMarkdown.js';
-export type { MdastOptions } from './tokensToMyst.js';
 export { tokensToMyst } from './tokensToMyst.js';
 
 export { plugins };
